some refactoring of the core files
