+++ blog/an_opinionated_set_of_testing_patterns/birthday_kata/app/src/main/kotlin/birthday_kata/core/domain/EmailClient.kt
@@ -1,0 +1,12 @@
+package birthday_kata.core.domain
+
+import arrow.core.Either
+
+sealed class SendError {
+    data class ConnectionFailed(val message: String): SendError()
+    data class AddressDoesNotExist(val emailAddress: EmailAddress): SendError()
+}
+interface EmailClient {
+    suspend fun send(email: Email): Either<SendError, Unit>
+}
+
