+++ blog/an_opinionated_set_of_testing_patterns/birthday_kata/app/src/main/kotlin/birthday_kata/core/domain/Employee.kt
@@ -1,0 +1,27 @@
+package birthday_kata.core.domain
+
+import java.time.LocalDate
+import java.time.MonthDay
+import java.util.UUID
+
+typealias Employees = List<Employee>
+
+data class Employee(
+    val id: EmployeeId,
+    val dateOfBirth: LocalDate,
+    val firstName: FirstName,
+    val lastName: LastName,
+    val emailAddress: EmailAddress
+)
+
+typealias FirstName = String
+typealias LastName = String
+typealias EmployeeId = UUID
+typealias Birthday = MonthDay
+
+fun Employee.toBirthdayEmail(): Email =
+    Email(
+        subject = "Happy Birthday!",
+        to = this.emailAddress,
+        body = "Dear ${this.firstName}, Happy Birthday!"
+    )
