+++ blog/an_opinionated_set_of_testing_patterns/birthday_kata/app/src/main/kotlin/birthday_kata/core/domain/EmployeeRepo.kt
@@ -1,0 +1,14 @@
+package birthday_kata.core.domain
+
+import arrow.core.Either
+
+sealed class FindByBirthdayError {
+    data class ConnectionFailed(val message: String): FindByBirthdayError()
+}
+sealed class SaveError {
+    data class ConnectionFailed(val message: String): SaveError()
+}
+interface EmployeeRepo {
+    suspend fun findByBirthday(dob: Birthday): Either<FindByBirthdayError, List<Employee>>
+    suspend fun save(employee: Employee): Either<SaveError, Unit>
+}
