+++ blog/an_opinionated_set_of_testing_patterns/birthday_kata/app/src/test/kotlin/birthday_kata/ConnectionFailureEmployeeRepo.kt
@@ -2,13 +2,11 @@ package birthday_kata
 
 import arrow.core.Either
 import arrow.core.left
-import birthday_kata.core.Birthday
-import birthday_kata.core.Employee
-import birthday_kata.core.EmployeeRepo
-import birthday_kata.core.FindByBirthdayError
-import birthday_kata.core.SaveError
-import java.time.LocalDate
-import java.time.MonthDay
+import birthday_kata.core.domain.Birthday
+import birthday_kata.core.domain.Employee
+import birthday_kata.core.domain.EmployeeRepo
+import birthday_kata.core.domain.FindByBirthdayError
+import birthday_kata.core.domain.SaveError
 
 object ConnectionFailureEmployeeRepo: EmployeeRepo {
     override suspend fun findByBirthday(dob: Birthday): Either<FindByBirthdayError, List<Employee>> =
@@ -17,6 +15,3 @@ object ConnectionFailureEmployeeRepo: EmployeeRepo {
     override suspend fun save(employee: Employee): Either<SaveError, Unit> =
         SaveError.ConnectionFailed(message="Oops!").left()
 }
-
-private fun LocalDate.matches(monthDay: MonthDay): Boolean =
-    MonthDay.from(this) == monthDay
