+++ blog/an_opinionated_set_of_testing_patterns/birthday_kata/app/src/test/kotlin/birthday_kata/StubbedEmailClient.kt
@@ -2,9 +2,9 @@ package birthday_kata
 
 import arrow.core.Either
 import arrow.core.right
-import birthday_kata.core.Email
-import birthday_kata.core.EmailClient
-import birthday_kata.core.SendError
+import birthday_kata.core.domain.Email
+import birthday_kata.core.domain.EmailClient
+import birthday_kata.core.domain.SendError
 
 object StubbedEmailClient: EmailClient {
     override suspend fun send(email: Email): Either<SendError, Unit> =
