+++ blog/an_opinionated_set_of_testing_patterns/birthday_kata/app/src/test/kotlin/birthday_kata/specs/EmailTest.kt
@@ -1,8 +1,9 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package birthday_kata
+package birthday_kata.specs
 
+import birthday_kata.Given
+import birthday_kata.`then Doug and Trixie should receive emails`
+import birthday_kata.`then no one should receive an email`
+import birthday_kata.`then only Doug should receive an email`
 import io.kotest.core.spec.style.StringSpec
 
 class EmailTest : StringSpec({
