+++ blog/an_opinionated_set_of_testing_patterns/birthday_kata/app/src/test/kotlin/birthday_kata/specs/EmployeeRepoFailureTest.kt
@@ -1,5 +1,7 @@
-package birthday_kata
+package birthday_kata.specs
 
+import birthday_kata.Given
+import birthday_kata.`then an employee lookup failed error should be returned`
 import io.kotest.core.spec.style.StringSpec
 
 class EmployeeRepoFailureTest : StringSpec({
